requirements.txt
djangorestframework
markdown
django-filter
django-rest-durin
django-email-verification
